move FileEditSource to event.py
