--- openhands/events/common.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class FileEditSource(str, Enum):
-    LLM_BASED_EDIT = 'llm_based_edit'
-    OH_ACI = 'oh_aci'
