+++ openhands/events/event.py
@@ -12,6 +12,11 @@ class EventSource(str, Enum):
     ENVIRONMENT = 'environment'
 
 
+class FileEditSource(str, Enum):
+    LLM_BASED_EDIT = 'llm_based_edit'
+    OH_ACI = 'oh_aci'
+
+
 @dataclass
 class Event:
     @property
